Moved Gamma.h/.cpp into Win32/src folder. Updated MSVC projects. Corrected pch.h capitalisation of Gamma.h
