+++ win32/src/Gamma.cpp
