+++ win32/src/Gamma.h
