+++ win32/src/pch.h
@@ -15,7 +15,7 @@
 #include "fixed.h"
 #include "Frame.h"
 #include "gameconsts.h"
-#include "gamma.h"
+#include "Gamma.h"
 #include "GeoSphere.h"
 //#include "glfreetype.h" 
 #include "gui/Gui.h"
