+++ win32/vc2010/pioneer.vcxproj.filters
@@ -474,8 +474,8 @@
     <ClCompile Include="..\..\src\render\RenderFrustum.cpp">
       <Filter>src</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\Gamma.cpp">
-      <Filter>src</Filter>
+    <ClCompile Include="..\src\Gamma.cpp">
+      <Filter>win32\src</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -992,8 +992,8 @@
     <ClInclude Include="..\..\src\render\RenderFrustum.h">
       <Filter>src</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\Gamma.h">
-      <Filter>src</Filter>
+    <ClInclude Include="..\src\Gamma.h">
+      <Filter>win32\src</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
